Move spring's kotlin processors and substitutions to a dedicated file
As the number of kotlin related substitutions and processing logic
seems to grow, it is better to move them to a separate file/class
to make it easier to locate kotlin specific logic.
